+++ Cluedo/src/cluedo/render/Screen.java
@@ -24,9 +24,9 @@ public class Screen extends Canvas{
 	private PlayerMouse mouse;
 
 	public Screen(int width, int height){
-		Dimension size = new Dimension(width,height);	
+		Dimension size = new Dimension(width,height);
 		System.out.println("Dimension: "+width +" "+height);
-		setSize(size);		
+		setSize(size);
 		setFocusable(true);
 		setBackground(new Color(200,200,200));
 		key = new PlayerKey();
@@ -34,15 +34,15 @@ public class Screen extends Canvas{
 		addKeyListener(key);
 		addMouseListener(mouse);
 	}
-	
+
 	public PlayerKey get_key(){
 		return this.key;
 	}
-	
+
 	public PlayerMouse get_mouse(){
 		return this.mouse;
 	}
-	
+
 	public void createBStrategy(){
 		BufferStrategy bs = getBufferStrategy();
 		if (bs == null) {
@@ -58,46 +58,46 @@ public class Screen extends Canvas{
 //			// This sets the buffering to do Double buffering
 //			createBufferStrategy(2);
 //			//return;
-//		}	
-		
-		Graphics2D g = (Graphics2D)bs.getDrawGraphics();	
+//		}
+
+		Graphics2D g = (Graphics2D)bs.getDrawGraphics();
 		//System.out.println("here");
-		render_map(g, as);	
+		render_map(g, as);
 		render_grid(g, as);
 		render_chars(g,as);
 		render_path(g,as);
 		bs.show();
 	}
-	
+
 	private void render_grid(Graphics2D g, Data as){
 		Tile[][] tiles = as.getTiles();
-		
+
 		for(int i = 0; i < tiles.length; i++){
 			for(int j = 0; j < tiles[0].length; j++){
 				g.setColor(Color.red);
 				Tile tile = tiles[i][j];
 				Rectangle2D r2 = new Rectangle2D.Double(tile.getPos().getX(),tile.getPos().getY(),tile.getTileSize() , tile.getTileSize());
-				g.draw(r2);					
+				g.draw(r2);
 			}
 		}
-		
+
 	}
 	private void render_path(Graphics2D g, Data da){
 		ArrayList<Point> points = da.getMousePath();
 		Tile[][] tiles = da.getTiles();
-		int pSize = da.getTileSize()/2;
+		double pSize = da.getTileSize()/2;
 		if(points.size()>0){
 			for(int i = 0; i < points.size();i++){
 				Point point = points.get(i);
 				Tile tile = tiles[point.y][point.x];
 				double ox = tile.getPos().getX();
 				double oy = tile.getPos().getY();
-				double p = da.getTileSize()/2-pSize/2;				
+				double p = da.getTileSize()/2-pSize/2;
 				g.setColor(Color.GREEN);
 				Rectangle2D r2 = new Rectangle2D.Double(p+ox,p+oy,pSize,pSize);
-				g.fill(r2);				
+				g.fill(r2);
 			}
-		}		
+		}
 	}
 
 	private void render_map(Graphics2D g, Data as){
@@ -106,16 +106,16 @@ public class Screen extends Canvas{
 		double h = getHeight();
 		double width_ratio = w/img.getWidth();
 		double height_ration = h/img.getHeight();
-		double ratio = Math.min(width_ratio,height_ration);			
-		
+		double ratio = Math.min(width_ratio,height_ration);
+
 		int image_width = (int) Math.round(img.getWidth()*ratio);
 		int image_height = (int) Math.round(img.getHeight()*ratio);
 		int pos_x = (int)Math.round((getWidth()>>1)-(image_width>>1));
-		int pos_y = (int)Math.round((getHeight()>>1)-(image_height>>1));		
-		
-		g.drawImage(img, pos_x, pos_y, image_width, image_height, null);		
+		int pos_y = (int)Math.round((getHeight()>>1)-(image_height>>1));
+
+		g.drawImage(img, pos_x, pos_y, image_width, image_height, null);
 	}
-	
+
 	private void render_chars(Graphics2D g, Data as){
 		ArrayList<Chars> chars = as.getAllChars();
 		Tile[][] tiles = as.getTiles();
